requirements.txt
# Lightweight requirements for Gemini Vision API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.1.0
